fix(server.ts): Change port variable case from lowercase port to uppercase PORT
The port variable in the server.ts file was changed from lowercase "port" to uppercase "PORT". This change was made to improve the semantics of the code and make it more readable.

feat(server.ts): Add support for process.env.PORT environment variable

In the server.ts file, support for the process.env.PORT environment variable was added. This allows the application to be run on a configurable port specified by the environment variable. This change was made to provide flexibility in running the application on different ports based on the deployment environment.